requirements.txt
langchain_community.llms 
langchain_core.prompts 
langchain_core.output_parsers 
streamlit 
langchain
